requirements.txt
# Scraping
requests>=2.31.0
beautifulsoup4>=4.12.2

# LLM et RAG
langchain>=0.0.267
langchain-core>=0.1.7
langchain-community>=0.0.11
langchain-text-splitters>=0.0.1
transformers>=4.33.2
sentence-transformers>=2.2.2
chromadb>=0.4.15

# Embeddings et modèles
huggingface-hub>=0.19.0
accelerate>=0.23.0
llama-cpp-python>=0.2.0

# NLP et traitement de texte
nltk>=3.8.1

# Interface Web
flask>=2.3.3

# Utilitaires
tqdm>=4.66.1